First Iteration
The first iteration of the model is complete as described in
Poort et al. 2012. A parameter file can be read in and on each
iteration every stage feeds activity forward and backwards when
possible. The lowest possible layer is fed an image and the
differential equations define how the cells update and interact.
Images produced in the notebook resemble those in the paper.