Model Version 2
Border detection and filtering of signals have been updated to use
a gaussian kernel which is convolved with the activity.
The order of operations has changed to reflect the order used in
Poort et al. Some data collection is now performed by the model
but this will most likely change.